:whit_check_mark: Test: HealthSection test
